# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Cooperative Services API",
    "description" : "Runnable prototype. Development OTP and simulated payments are explicitly gated. See docs/api-contract.md for response examples and integration rules.",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/admin/workers/{id}/verification" : {
      "put" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "verify",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Verification"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/admin/workers/{id}/insurance" : {
      "put" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "insurance",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Insurance"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/admin/subservices/{id}/price" : {
      "put" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "price",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Price"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/admin/config" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "config",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "config_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Config"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/jobs/{id}/payment" : {
      "get" : {
        "tags" : [ "billing-controller" ],
        "operationId" : "payment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/catalog/subservices" : {
      "get" : {
        "tags" : [ "catalog-controller" ],
        "operationId" : "services",
        "parameters" : [ {
          "name" : "categoryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/workers" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "workers",
        "parameters" : [ {
          "name" : "societyId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/admin/welfare" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "welfare",
        "parameters" : [ {
          "name" : "societyId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/admin/unfulfilled" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "unfulfilled",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/admin/jobs/{id}/eligible-workers" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "eligibleWorkers",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/admin/jobs/{id}/allocation" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "allocation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/admin/audit" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "audit",
        "parameters" : [ {
          "name" : "after",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "default" : 0,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Verification" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "pattern" : "ACTIVE|REJECTED|SUSPENDED"
          },
          "verifiedCategoryIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "uniqueItems" : true
          },
          "note" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          }
        },
        "required" : [ "note", "status", "verifiedCategoryIds" ]
      },
      "Insurance" : {
        "type" : "object",
        "properties" : {
          "pmsbyStatus" : {
            "type" : "string",
            "pattern" : "NOT_ENROLLED|PENDING|ENROLLED"
          },
          "pmjjbyStatus" : {
            "type" : "string",
            "pattern" : "NOT_ENROLLED|PENDING|ENROLLED"
          },
          "evidenceReference" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          }
        },
        "required" : [ "evidenceReference", "pmjjbyStatus", "pmsbyStatus" ]
      },
      "Price" : {
        "type" : "object",
        "properties" : {
          "basePrice" : {
            "type" : "number",
            "minimum" : 0.01
          },
          "emergencySupported" : {
            "type" : "boolean"
          },
          "active" : {
            "type" : "boolean"
          }
        },
        "required" : [ "basePrice" ]
      },
      "Config" : {
        "type" : "object",
        "properties" : {
          "version" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "proximityWeight" : {
            "type" : "number",
            "maximum" : 1,
            "minimum" : 0
          },
          "ratingWeight" : {
            "type" : "number",
            "maximum" : 1,
            "minimum" : 0
          },
          "loadWeight" : {
            "type" : "number",
            "maximum" : 1,
            "minimum" : 0
          },
          "welfareRate" : {
            "type" : "number",
            "maximum" : 1,
            "minimum" : 0
          },
          "emergencySurcharge" : {
            "type" : "number",
            "minimum" : 0
          },
          "standardRadiusM" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 50000,
            "minimum" : 100
          },
          "emergencyRadiusM" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 20000,
            "minimum" : 100
          },
          "emergencyTimeoutS" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 90,
            "minimum" : 60
          }
        },
        "required" : [ "emergencySurcharge", "loadWeight", "proximityWeight", "ratingWeight", "welfareRate" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}